fix broken cmap output caused by move to char32_t

+++ source/pdf/font/cmap.cpp
@@ -19,9 +19,10 @@ namespace pdf
 		auto cmap = this->unicode_cmap;
 		write_cmap_header(cmap, this->pdf_font_name);
 
-		cmap->append("1 begincodespacerange\n"
-					 "<0000> <FFFF>\n"
-					 "endcodespacerange\n");
+		cmap->append(
+			"1 begincodespacerange\n"
+			"<0000> <FFFF>\n"
+			"endcodespacerange\n");
 
 		assert(this->source_file != nullptr);
 		auto& mapping = this->source_file->character_mapping.forward;
@@ -54,12 +55,12 @@ namespace pdf
 		cmap->append(zpr::sprint("{} beginbfchar\n", m_extra_unicode_mappings.size()));
 		for(auto& [gid, cps] : m_extra_unicode_mappings)
 		{
-			cmap->append(zpr::sprint("<{04x}> <", gid));
+			cmap->append(zpr::sprint("<{04x}> <", static_cast<uint32_t>(gid)));
 			for(auto cp : cps)
 			{
 				if(cp <= 0xFFFF)
 				{
-					cmap->append(zpr::sprint("{04x}", cp));
+					cmap->append(zpr::sprint("{04x}", static_cast<uint16_t>(cp)));
 				}
 				else
 				{
@@ -90,23 +91,24 @@ namespace pdf
 		auto registry = "Sap";
 		auto supplement = 0;
 
-		auto header = zpr::sprint("%!PS-Adobe-3.0 Resource-CMap\n"
-								  "%%DocumentNeededResources: ProcSet (CIDInit)\n"
-								  "%%IncludeResource: ProcSet (CIDInit)\n"
-								  "%%BeginResource: CMap ({})\n"
-								  "%%Title: ({} {} {} {})\n"
-								  "%%Version: 1\n"
-								  "%%EndComments\n"
-								  "/CIDInit /ProcSet findresource begin\n"
-								  "12 dict begin\n"
-								  "begincmap\n"
-								  "/CIDSystemInfo <<\n"
-								  "  /Registry ({})\n"
-								  "  /Ordering ({})\n"
-								  "  /Supplement {}\n"
-								  ">> def\n"
-								  "/CMapName /{} def\n"
-								  "/CMapType 2 def\n",
+		auto header = zpr::sprint(
+			"%!PS-Adobe-3.0 Resource-CMap\n"
+			"%%DocumentNeededResources: ProcSet (CIDInit)\n"
+			"%%IncludeResource: ProcSet (CIDInit)\n"
+			"%%BeginResource: CMap ({})\n"
+			"%%Title: ({} {} {} {})\n"
+			"%%Version: 1\n"
+			"%%EndComments\n"
+			"/CIDInit /ProcSet findresource begin\n"
+			"12 dict begin\n"
+			"begincmap\n"
+			"/CIDSystemInfo <<\n"
+			"  /Registry ({})\n"
+			"  /Ordering ({})\n"
+			"  /Supplement {}\n"
+			">> def\n"
+			"/CMapName /{} def\n"
+			"/CMapType 2 def\n",
 
 			cmap_name, cmap_name, registry, ordering, supplement, registry, ordering, supplement, cmap_name);
 
@@ -115,11 +117,12 @@ namespace pdf
 
 	static void write_cmap_footer(Stream* s)
 	{
-		s->append(zst::str_view("endcmap\n"
-								"CMapName currentdict /CMap defineresource pop\n"
-								"end\n"
-								"end\n"
-								"%%EndResource"
-								"%%EOF"));
+		s->append(zst::str_view(
+			"endcmap\n"
+			"CMapName currentdict /CMap defineresource pop\n"
+			"end\n"
+			"end\n"
+			"%%EndResource"
+			"%%EOF"));
 	}
 }
